Move sendActivationEmail to user/lib
